+++ .gitignore
@@ -62,3 +62,5 @@ typings/
 
 # Internal dir
 internal/
+
+dist/
