+++ jest.config.js
@@ -1,0 +1,13 @@
+/** @type {import('ts-jest').JestConfigWithTsJest} */
+export default {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  testMatch: ['<rootDir>/**/tests/**/*.test.ts'],
+  testPathIgnorePatterns: ['/node_modules/'],
+  coverageDirectory: './coverage',
+  coveragePathIgnorePatterns: ['node_modules'],
+  reporters: ['default'],
+  transform: {
+    '^.+\\.tsx?$': ['ts-jest']
+  }
+};
