+++ package.json
@@ -7,7 +7,9 @@
   "author": "Saektide <real.saektide@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "test": "jest --verbose"
+    "test": "jest --verbose",
+    "web": "cd playground && pnpm dev",
+    "build": "tsc && tsc-alias"
   },
   "keywords": [
     "dtext",
@@ -15,12 +17,15 @@
     "e621"
   ],
   "devDependencies": {
-    "jest": "^26.6.3"
-  },
-  "jest": {
-    "testTimeout": 20000
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.11.25",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.2",
+    "tsc-alias": "^1.8.8",
+    "typescript": "^5.4.2"
   },
   "engines": {
     "node": ">=12.13.0"
-  }
+  },
+  "type": "module"
 }
