+++ tests/parserFormats.test.ts
@@ -1,8 +1,8 @@
-const dtext = require('../');
+import dtext from '../src';
 
 test('parse format Bold', () => {
   return dtext.parse('[b]Test[/b]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<b>Test</b>');
     });
@@ -10,7 +10,7 @@ test('parse format Bold', () => {
 
 test('parse format Italics', () => {
   return dtext.parse('[i]Test[/i]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<i>Test</i>');
     });
@@ -18,7 +18,7 @@ test('parse format Italics', () => {
 
 test('parse format Underline', () => {
   return dtext.parse('[u]Test[/u]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<u>Test</u>');
     });
@@ -26,7 +26,7 @@ test('parse format Underline', () => {
 
 test('parse format Overline', () => {
   return dtext.parse('[o]Test[/o]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<o>Test</o>');
     });
@@ -34,7 +34,7 @@ test('parse format Overline', () => {
 
 test('parse format Strikeout', () => {
   return dtext.parse('[s]Test[/s]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<s>Test</s>');
     });
@@ -42,7 +42,7 @@ test('parse format Strikeout', () => {
 
 test('parse format Superscript', () => {
   return dtext.parse('[sup]Test[/sup]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<sup>Test</sup>');
     });
@@ -50,7 +50,7 @@ test('parse format Superscript', () => {
 
 test('parse format Subscript', () => {
   return dtext.parse('[sub]Test[/sub]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<sub>Test</sub>');
     });
@@ -58,7 +58,7 @@ test('parse format Subscript', () => {
 
 test('parse format Username', () => {
   return dtext.parse('@Test')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<b>@Test</b>');
     });
@@ -66,7 +66,7 @@ test('parse format Username', () => {
 
 test('parse format Spoiler', () => {
   return dtext.parse('[spoiler]Test[/spoiler]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<span class="spoiler">Test</span>');
     });
@@ -74,7 +74,7 @@ test('parse format Spoiler', () => {
 
 test('parse format Code', () => {
   return dtext.parse('`Test`')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<code>Test</code>');
     });
@@ -82,7 +82,7 @@ test('parse format Code', () => {
 
 test('parse format Inline code', () => {
   return dtext.parse('`Test`')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<code>Test</code>');
     });
@@ -90,7 +90,7 @@ test('parse format Inline code', () => {
 
 test('parse format Color', () => {
   return dtext.parse('[color=#f00]Test[/color]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<span style="color:#f00">Test</span>');
     });
@@ -98,7 +98,7 @@ test('parse format Color', () => {
 
 test('parse format External link', () => {
   return dtext.parse('https://google.com')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<a href="https://google.com">https://google.com</a>');
     });
@@ -106,7 +106,7 @@ test('parse format External link', () => {
 
 test('parse format Wiki link', () => {
   return dtext.parse('[[test]]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<a href="https://e621.net/wiki_pages/test">test</a>');
     });
@@ -114,7 +114,7 @@ test('parse format Wiki link', () => {
 
 test('parse format Tag link', () => {
   return dtext.parse('{{test}}')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<a href="https://e621.net/posts?tags=test">test</a>');
     });
@@ -122,7 +122,7 @@ test('parse format Tag link', () => {
 
 test('parse format Intern link', () => {
   return dtext.parse('post #1234')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<a href="https://e621.net/posts/1234">post #1234</a>');
     });
@@ -130,7 +130,7 @@ test('parse format Intern link', () => {
 
 test('parse format Quote', () => {
   return dtext.parse('[quote]Test[/quote]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<blockquote>Test</blockquote>');
     });
@@ -138,7 +138,7 @@ test('parse format Quote', () => {
 
 test('parse format Code', () => {
   return dtext.parse('[code]Test[/code]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<pre>Test</pre>');
     });
@@ -146,7 +146,7 @@ test('parse format Code', () => {
 
 test('parse format Header', () => {
   return dtext.parse('h1.Test')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<h1>Test</h1>');
     });
@@ -154,7 +154,7 @@ test('parse format Header', () => {
 
 test('parse format Section', () => {
   return dtext.parse('[section,expanded="Test"]Test[/section]')
-    .then((result) => {
+    .then((result:string) => {
       expect(result).toBe
         ('<section class="section collapsed"><div class="section-collapse-title">"Test"]Test[/section</div><div class="section-collapsed-title">"Test"]Test[/section</div><div class="section-content">');
     });
